简单的完成了磁盘调度算法的模拟实现与显示

+++ widget.cpp
@@ -14,13 +14,23 @@ Widget::Widget(QWidget *parent) :
     {
         emptyBlockList.addEmptyBlock(i);
     }
-
+    //初始磁盘调度算法数据结构
+    waitIDByVector.resize(trackNumber);
     int blockID=emptyBlockList.getEmptyBlock();
     disk[blockID].setFlag(DiskBlock::DIR);
+    //初始化计时器
+    updateTimer= new QTimer(this);
+    //设置定时器循环触发
+    updateTimer->setSingleShot(false);
+    //触发时间间隔为1s
+    updateTimer->setInterval(1000);
+    connect(updateTimer,&QTimer::timeout,this,&Widget::updateGUI);
+    updateTimer->start();
 }
 
 Widget::~Widget()
 {
+    delete updateTimer;
     delete ui;
 }
 
@@ -50,11 +60,13 @@ void Widget::on_open_clicked()
             return;
         }
         id=diskID;
+        addWaitBlock(id);
     }
     inodeID=id;
     ui->fileName->setText(nameList.last());
     QString content;
     while (id>=0) {
+        addWaitBlock(id);
         content+=disk[id].getFileContent();
         id=disk[id].getNextBlock();
     }
@@ -98,6 +110,7 @@ void Widget::on_newFile_clicked()
             {
                 item=findItem(item,nameList[i]);
                 id=diskID;
+                addWaitBlock(id);
             }
             else
             {
@@ -106,6 +119,7 @@ void Widget::on_newFile_clicked()
                 {
                     item=findItem(item,nameList[i]);
                     id=diskID;
+                    addWaitBlock(id);
                 }
 
             }
@@ -125,6 +139,7 @@ void Widget::on_newFile_clicked()
             {
                 item=findItem(item,nameList[i]);
                 id=diskID;
+                addWaitBlock(id);
             }
             else
             {
@@ -133,12 +148,15 @@ void Widget::on_newFile_clicked()
                 {
                     item=findItem(item,nameList[i]);
                     id=diskID;
+                    addWaitBlock(id);
                 }
 
             }
         }
 
-        disk[id].newFile(nameList[nameList.size()-1],emptyBlockList,disk,item);
+        id=disk[id].newFile(nameList[nameList.size()-1],emptyBlockList,disk,item);
+        if(id>=0)
+            addWaitBlock(id);
     }
 }
 
@@ -150,9 +168,11 @@ void Widget::on_save_clicked()
         //一个字符一个字节，一个block可以存放128个字符
         if(content.size()<=128)
         {
+            addWaitBlock(id);
             disk[id].setFileContent(content);
             break;
         }else {
+            addWaitBlock(id);
             disk[id].setFileContent(content.mid(128));
             content=content.mid(128,content.size());
         }
@@ -170,6 +190,7 @@ void Widget::on_save_clicked()
     //释放可能存在的剩余节点
     id=disk[id].getNextBlock();
     while (id>=0) {
+        addWaitBlock(id);
         int nextBlock=disk[id].getNextBlock();
         disk[id].clear();
         emptyBlockList.addEmptyBlock(id);
@@ -177,23 +198,26 @@ void Widget::on_save_clicked()
     }
 }
 
-void Widget::deleteFile(int id,QTreeWidgetItem * item)
+int Widget::deleteFile(int id,QTreeWidgetItem * item)
 {
     if(id==inodeID)
     {
+        addWaitBlock(id);
         QString fileName=item->text(0);
         QMessageBox::warning(NULL,"警告",tr("文件 %1 正在被编辑无法删除").arg(fileName));
-        return;
+        return -1;
     }
     //删除文件的树形控件
     delete item;
     //释放磁盘空间
     while (id>=0) {
+        addWaitBlock(id);
         int nextBlock=disk[id].getNextBlock();
         disk[id].clear();
         emptyBlockList.addEmptyBlock(id);
         id=nextBlock;
     }
+    return 0;
 }
 
 void Widget::deleteDIR(int id,QTreeWidgetItem * treeItem)
@@ -214,6 +238,7 @@ void Widget::deleteDIR(int id,QTreeWidgetItem * treeItem)
                 deleteDIR(item.second,tempItem);
             }
         }
+        addWaitBlock(id);
         disk[id].clear();
         emptyBlockList.addEmptyBlock(id);
         id=nextBlock;
@@ -244,12 +269,14 @@ void Widget::on_deleteFile_clicked()
             {
                 item=findItem(item,nameList[i]);
                 id=diskID;
+                addWaitBlock(id);
             }else
             {
                 QMessageBox::critical(NULL,"警告","你输入的路径有误");
                 return;
             }
         }
+        deleteDIR(id,item);
 
     }else
     {
@@ -267,6 +294,7 @@ void Widget::on_deleteFile_clicked()
             {
                 item=findItem(item,nameList[i]);
                 id=diskID;
+                addWaitBlock(id);
             }else
             {
                 QMessageBox::critical(NULL,"警告","你输入的路径有误");
@@ -278,8 +306,8 @@ void Widget::on_deleteFile_clicked()
         int diskID=disk[id].fileExist(nameList.last(),disk);
         if(diskID>=0)
         {
-            deleteFile(diskID,item);
-            disk[id].deleteRecord(nameList.last());
+            if(deleteFile(diskID,item)>=0)
+                disk[id].deleteRecord(nameList.last());
         }else
         {
             QMessageBox::critical(NULL,"警告","你输入的路径有误");
@@ -292,4 +320,108 @@ void Widget::on_close_clicked()
     inodeID=-1;
     ui->fileName->setText("没有打开文件");
     ui->fileContent->setText("");
+}
+
+void Widget::updateGUI()
+{
+    if(waitIDByList.size()==0)
+        return;
+    if(ui->diskOption->currentText()=="FCFS")
+    {
+        int id=waitIDByList.front();
+        waitIDByList.pop_front();
+        int trackID=id%blockNumber;
+        //同步清除waitIDByVector的数据
+        int loopSize=waitIDByVector[trackID].size();
+        for(int i=0;i<loopSize;++i)
+        {
+            if(waitIDByVector[trackID][i]==id)
+            {
+                waitIDByVector[trackID].erase(waitIDByVector[trackID].begin()+i);
+                break;
+            }
+        }
+        ui->blockNumber->setText(QString::number(id));
+        ui->trackNumber->setText(QString::number(trackID));
+    }else if(ui->diskOption->currentText()=="SCAN")
+    {
+        int blockID=0;
+        int addNumber=0;
+        if(direction)
+            addNumber=1;
+        else
+            addNumber=-1;
+        while (true) {
+            if(waitIDByVector[trackID].size()>0)
+            {
+                blockID=waitIDByVector[trackID][waitIDByVector[trackID].size()-1];
+                waitIDByVector[trackID].pop_back();
+                break;
+            }
+            //到底反向
+            if(trackID==trackEnd&&direction)
+            {
+                direction=false;
+                addNumber=-1;
+            }
+            if(trackID==trackBegin&&!direction)
+            {
+                direction=true;
+                addNumber=1;
+            }
+            trackID+=addNumber;
+        }
+
+        for(auto i=waitIDByList.begin();i!=waitIDByList.end();++i)
+        {
+            if(*i==blockID)
+            {
+                waitIDByList.erase(i);
+                break;
+            }
+        }
+        ui->blockNumber->setText(QString::number(blockID));
+        ui->trackNumber->setText(QString::number(trackID));
+    }
+    else if(ui->diskOption->currentText()=="CSCAN")
+    {
+        int blockID=0;
+        while (true)
+        {
+            if(waitIDByVector[trackID].size()>0)
+            {
+                blockID=waitIDByVector[trackID][waitIDByVector[trackID].size()-1];
+                waitIDByVector[trackID].pop_back();
+                break;
+            }
+            //到底反向
+            if(trackID==trackEnd&&waitIDByVector[trackID].size()==0)
+            {
+                trackID=trackBegin;
+            }
+            ++trackID;
+        }
+        for(auto i=waitIDByList.begin();i!=waitIDByList.end();++i)
+        {
+            if(*i==blockID)
+            {
+                waitIDByList.erase(i);
+                break;
+            }
+        }
+        ui->blockNumber->setText(QString::number(blockID));
+        ui->trackNumber->setText(QString::number(trackID));
+    }
+}
+
+void Widget::addWaitBlock(int id)
+{
+    //更新开始结束标志
+    if(id>trackEnd)
+        trackEnd=id;
+    if(id<trackBegin)
+        trackBegin=id;
+    waitIDByList.push_back(id);
+    int trackID=id%trackNumber;
+    waitIDByVector[trackID].push_back(id);
 }
